+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## Next release
+
+### ✨ Improved
+
+* Wrap reading TPM data in a try/except.
+
+
 ## 1.3.2 - July 10, 2025
 
 ### ✨ Improved